[fontbe] Move usMaxContext to own module
This is a bunch of code and we're going to add to it, so I think it
makes sense to have this live on its own.
